Remove non-static paragraph
The non-static builds are currently being removed as it's a pain to maintain them. They are substantially smaller than the fully static builds, but less robust and require building on a glibc distro, which would require 2 more containers in addition to the Alpine builds. I may eventually re-add them, but it won't be until I can at least make the shImg runtime more robust (eg: give proper errors if opened on a non-compatible distro)